refactor: extract CLI runtime command wrappers
